+++ hc/front/views.py
@@ -825,7 +825,7 @@ def details(request, code):
         "channels": channels,
         "enabled_channels": list(check.channel_set.all()),
         "timezones": all_timezones,
-        "downtimes": check.downtimes(months=3),
+        "downtimes": check.downtimes(3, request.profile.tz),
         "is_copied": "copied" in request.GET,
         "all_tags": " ".join(sorted(all_tags)),
     }
@@ -922,7 +922,8 @@ def status_single(request, code):
 
     if updated != request.GET.get("u"):
         doc["events"] = EVENTS_TMPL.render({"check": check, "events": events})
-        doc["downtimes"] = DOWNTIMES_TMPL.render({"downtimes": check.downtimes(3)})
+        downtimes = check.downtimes(3, request.profile.tz)
+        doc["downtimes"] = DOWNTIMES_TMPL.render({"downtimes": downtimes})
 
     return JsonResponse(doc)